# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Music Albums and Songs API",
      "version": "1.0.0",
      "description": "API for managing the Music Albums and Songs dataset.",
      "contact": {
        "name": "Vito Vrbić",
        "email": "vito.vrbic@fer.unizg.hr"
      },
      "license": {
        "name": "CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
        "url": "https://creativecommons.org/publicdomain/zero/1.0/"
      }
    },
    "paths": {
...
      "/albums/{id}": {
        "get": {
          "summary": "Get an album by ID with detailed information",
          "description": "Fetches a single album by its ID, including details about songs, producers, and performers.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the album",
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Album details with songs, producers, and performers",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Album"
                  }
                }
              }
            },
            "404": {
              "description": "Album not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update an album by ID",
          "description": "Update the details of an existing album by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the album to be updated",
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Album updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Album"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid data provided"
            },
            "404": {
              "description": "Album not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete an album by ID",
          "description": "Deletes an album by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the album to be deleted",
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Album deleted successfully"
            },
            "404": {
              "description": "Album not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Album": {
          "type": "object",
          "properties": {
            "album_id": {
              "type": "integer",
              "format": "int64",
              "description": "Unique identifier of the album"
            },
            "album_title": {
              "type": "string",
              "description": "Title of the album"
            },
            "release_year": {
              "type": "integer",
              "format": "int64",
              "description": "Year the album was released"
            },
            "number_of_singles": {
              "type": "integer",
              "format": "int64",
              "description": "Number of singles released from the album"
            },
            "genre": {
              "type": "string",
              "description": "Genre of the album"
            },
            "style": {
              "type": "string",
              "description": "Style of the album"
            },
            "release_label": {
              "type": "string",
              "description": "Record label of the album"
            },
            "country": {
              "type": "string",
              "description": "Country where the album was released"
            },
            "type": {
              "type": "string",
              "description": "Type of album (e.g., studio, live)"
            },
            "songs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Song"
              }
            },
            "producers": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Producer"
              }
            },
            "performers": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Performer"
              }
            }
          }
        },
        "Song": {
          "type": "object",
          "properties": {
            "song_title": {
              "type": "string",
              "description": "Title of the song"
            },
            "track_number": {
              "type": "integer",
              "description": "Track number of the song in the album"
            },
            "duration": {
              "type": "string",
              "description": "Duration of the song"
            }
          }
        },
        "Producer": {
          "type": "object",
          "properties": {
            "artist_name": {
              "type": "string",
              "description": "Name of the producer"
            }
          }
        },
        "Performer": {
          "type": "object",
          "properties": {
            "artist_name": {
              "type": "string",
              "description": "Name of the performer"
            }
          }
        }
      }
    }
  }